fix migration

+++ cncnet-api/database/migrations/2022_11_03_224557_CanceledMatches.php
@@ -15,8 +15,7 @@ class CanceledMatches extends Migration
         Schema::create('qm_canceled_matches', function (Blueprint $table)
 		{
 			$table->increments('id');
-
-			$table->integer('qm_match_id')->unsigned();
+			$table->unsignedBigInteger('qm_match_id')->unsigned();
 			$table->integer('player_id')->unsigned();
             $table->integer('ladder_id')->unsigned();
